# contracts/openapi/m-ranked-v1.yaml
openapi: 3.1.0
info:
  title: M-Ranked API
  version: 1.0.0
  description: >-
    Public reads are backed by revision-pinned projections and legacy identifier aliases.
    Administrative status and account commands are authenticated, role-scoped and never cached.
...
paths:
...
  /api/v1/admin/csrf:
    get:
      tags: [Admin]
      operationId: getAdminCsrf
      summary: Issue a double-submit CSRF token for an authenticated admin client
      security:
        - basicAuth: []
      responses:
        '200':
          description: Token metadata; the matching XSRF-TOKEN cookie is also issued.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminCsrf'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/admin/jobs:
    get:
      tags: [Admin]
      operationId: getAdminJobs
      summary: Read bounded collection-run status
      description: Requires VIEWER, EDITOR or ADMIN. No raw response or payload is returned.
      security:
        - basicAuth: []
      parameters:
        - name: platform
          in: query
          schema:
            type: string
            enum: [telegram, vk, max, rutube]
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/RunStatus'
        - name: limit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50
      responses:
        '200':
          description: Newest collection runs matching the allowlisted filters.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminJobPage'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/AdminUnavailable'
  /api/v1/admin/jobs/{jobId}:
    get:
      tags: [Admin]
      operationId: getAdminJob
      summary: Read one collection run and bounded per-account outcomes
      description: Requires VIEWER, EDITOR or ADMIN. Errors are limited to sanitized codes.
      security:
        - basicAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: accountResultLimit
          in: query
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 100
      responses:
        '200':
          description: Collection run plus bounded account outcomes.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminJobDetail'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/AdminUnavailable'
  /api/v1/admin/platform-accounts/{accountId}:
    get:
      tags: [Admin]
      operationId: getAdminPlatformAccount
      summary: Read the current administrative state of one platform account
      description: >-
        Requires VIEWER, EDITOR or ADMIN. Returns only the account UUID, platform, enabled flag,
        optimistic row version and update time from the dedicated administrative database
        connection. External identifiers, URLs, credentials and raw collector evidence are omitted.
      security:
        - basicAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Current state suitable for preparing an optimistic admin command.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPlatformAccountState'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/AdminUnavailable'
  /api/v1/admin/platform-accounts/{accountId}/enabled:
    put:
      tags: [Admin]
      operationId: setAdminPlatformAccountEnabled
      summary: Set the collector-enabled state of one platform account
      description: >-
        Requires EDITOR or ADMIN. The command uses optimistic locking and is idempotent when the
        requested desired state is already current. Success, replay, conflict and not-found outcomes
        are written to append-only audit with only enabled/rowVersion state. A changed command
        rebuilds and publishes its revision-pinned projections in the same database transaction.
      security:
        - basicAuth: []
      parameters:
        - name: accountId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-XSRF-TOKEN
          in: header
          required: true
          description: Token returned by getAdminCsrf; the matching XSRF-TOKEN cookie is required.
          schema:
            type: string
        - name: X-Correlation-Id
          in: header
          required: false
          description: Optional caller UUID; the server generates one when absent.
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSetEnabledRequest'
      responses:
        '200':
          description: Updated state or an idempotent no-op.
          headers:
            Cache-Control:
              $ref: '#/components/headers/NoStore'
            X-Correlation-Id:
              description: Correlation UUID used by audit and revision lineage.
              schema:
                type: string
                format: uuid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminSetEnabledResponse'
        '400':
          $ref: '#/components/responses/InvalidRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/OptimisticLockConflict'
        '503':
          $ref: '#/components/responses/AdminUnavailable'
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: >-
        Transitional stateless admin authentication. Store only bcrypt hashes in runtime
        configuration and expose this surface exclusively behind TLS. OIDC remains a later slice.
...
  headers:
...
    NoStore:
      description: Response must not be stored.
      schema:
        type: string
        const: no-store
  responses:
    InvalidRequest:
      description: A query, path parameter, or cursor is invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: No alias-backed resource exists at this dataset revision.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    Unauthorized:
      description: Authentication is missing or invalid.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
        WWW-Authenticate:
          description: Basic authentication challenge for the transitional admin surface.
          schema:
            type: string
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: The authenticated role cannot perform the requested operation or CSRF failed.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    OptimisticLockConflict:
      description: The supplied expected row version is stale.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    AdminUnavailable:
      description: The dedicated administrative database connection is disabled or unavailable.
      headers:
        Cache-Control:
          $ref: '#/components/headers/NoStore'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    AdminCsrf:
      type: object
      additionalProperties: false
      required: [headerName, parameterName, token]
      properties:
        headerName:
          type: string
          const: X-XSRF-TOKEN
        parameterName:
          type: string
          const: _csrf
        token:
          type: string
          minLength: 1
    RunStatus:
      type: string
      enum: [pending, running, succeeded, partial, failed, skipped, cancelled]
    AdminJob:
      type: object
      additionalProperties: false
      required:
        - jobId
        - kind
        - platform
        - scheduledAt
        - startedAt
        - completedAt
        - status
        - accountCount
        - errorCount
        - correlationId
      properties:
        jobId:
          type: string
          format: uuid
        kind:
          type: string
          const: collection
        platform:
          type: string
          enum: [telegram, vk, max, rutube]
        scheduledAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        status:
          $ref: '#/components/schemas/RunStatus'
        accountCount:
          type: integer
          format: int32
          minimum: 0
        errorCount:
          type: integer
          format: int32
          minimum: 0
        correlationId:
          type: string
          format: uuid
    AdminJobPage:
      type: object
      additionalProperties: false
      required: [items]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/AdminJob'
    AdminAccountResult:
      type: object
      additionalProperties: false
      required:
        - resultId
        - platformAccountId
        - startedAt
        - completedAt
        - status
        - discoveredCount
        - snapshotCount
        - sanitizedErrorCode
      properties:
        resultId:
          type: integer
          format: int64
          minimum: 1
        platformAccountId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: [string, 'null']
          format: date-time
        status:
          $ref: '#/components/schemas/RunStatus'
        discoveredCount:
          type: integer
          format: int32
          minimum: 0
        snapshotCount:
          type: integer
          format: int32
          minimum: 0
        sanitizedErrorCode:
          type: [string, 'null']
    AdminJobDetail:
      type: object
      additionalProperties: false
      required: [job, accountResults, accountResultsTruncated]
      properties:
        job:
          $ref: '#/components/schemas/AdminJob'
        accountResults:
          type: array
          maxItems: 200
          items:
            $ref: '#/components/schemas/AdminAccountResult'
        accountResultsTruncated:
          type: boolean
    AdminSetEnabledRequest:
      type: object
      additionalProperties: false
      required: [enabled, expectedRowVersion]
      properties:
        enabled:
          type: boolean
        expectedRowVersion:
          type: integer
          format: int64
          minimum: 0
    AdminPlatformAccountState:
      type: object
      additionalProperties: false
      required: [accountId, platform, enabled, rowVersion, updatedAt]
      properties:
        accountId:
          type: string
          format: uuid
        platform:
          type: string
          enum: [telegram, vk, max, rutube]
        enabled:
          type: boolean
        rowVersion:
          type: integer
          format: int64
          minimum: 0
        updatedAt:
          type: string
          format: date-time
    AdminSetEnabledResponse:
      type: object
      additionalProperties: false
      required: [account, changed, datasetRevision, correlationId, outcome]
      properties:
        account:
          $ref: '#/components/schemas/AdminPlatformAccountState'
        changed:
          type: boolean
        datasetRevision:
          type: [integer, 'null']
          format: int64
          minimum: 1
          description: New ready configuration revision; null for an idempotent no-op.
        correlationId:
          type: string
          format: uuid
        outcome:
          type: string
          enum: [updated, idempotent]
...
    Problem:
      type: object
      required: [type, title, status, detail, instance]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
          format: uri-reference